Revert "Задание Больше модулей"
This reverts commit d3cc0ea72fe31126f09e50a21863b90f0dfb0a8c.
