--- src/load.js
@@ -1,11 +1,0 @@
-'use strict';
-
-(function() {
-  define(function() {
-    return function(url, callbackFunction) {
-      var elem = document.createElement('script');
-      elem.src = url + '/?callback=' + callbackFunction;
-      document.body.appendChild(elem);
-    };
-  });
-})();
