+++ src/main.js
@@ -1,10 +1,5 @@
 'use strict';
 
-define(['./pictures', './review', './load', './resizer', './upload', ], function(picturesContainer, getPictureElement, load) {
-  load('http://localhost:1506/api/pictures', 'getPictures');
-  window.getPictures = function(data) {
-    data.forEach(function(picture) {
-      getPictureElement(picture, picturesContainer);
-    });
-  };
-});
+require('./resizer.js');
+require('./upload.js');
+require('./pictures.js');
